backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.6.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
PyJWT==2.8.0
bcrypt==4.2.0
aiohttp==3.9.5
starlette==0.37.2

